+++ src/components.d.ts
@@ -387,12 +387,17 @@ export namespace Components {
      * @cssprop --wcs-button-font-size-l - font-size for a size l button
      * @cssprop --wcs-button-padding-size-l - padding for a size l button
      * @cssprop --wcs-button-font-weight - font weight of a plain,stroked,clear button
+     * @cssprop --wcs-button-gap - gap between button content (text and icons)
      */
     interface WcsButton {
         /**
           * Specify whether the button is disabled or not.
          */
         "disabled": boolean;
+        /**
+          * Name of a wcs-mat-icon to display at the end (right) of the button.
+         */
+        "endIcon"?: string;
         /**
           * Set a URL to point to.<br/> If specified use a `a` tag instead of `btn`.
          */
@@ -418,6 +423,10 @@ export namespace Components {
           * Specify the size of the button.
          */
         "size": WcsButtonSize;
+        /**
+          * Name of a wcs-mat-icon to display at the start (left) of the button.
+         */
+        "startIcon"?: string;
         /**
           * Specifies where to open the linked document when using href (see prop above)<br/> Default '_self' will open the linked document in the same frame as it was clicked
          */
@@ -2972,6 +2981,7 @@ declare global {
      * @cssprop --wcs-button-font-size-l - font-size for a size l button
      * @cssprop --wcs-button-padding-size-l - padding for a size l button
      * @cssprop --wcs-button-font-weight - font weight of a plain,stroked,clear button
+     * @cssprop --wcs-button-gap - gap between button content (text and icons)
      */
     interface HTMLWcsButtonElement extends Components.WcsButton, HTMLStencilElement {
     }
@@ -5118,12 +5128,17 @@ declare namespace LocalJSX {
      * @cssprop --wcs-button-font-size-l - font-size for a size l button
      * @cssprop --wcs-button-padding-size-l - padding for a size l button
      * @cssprop --wcs-button-font-weight - font weight of a plain,stroked,clear button
+     * @cssprop --wcs-button-gap - gap between button content (text and icons)
      */
     interface WcsButton {
         /**
           * Specify whether the button is disabled or not.
          */
         "disabled"?: boolean;
+        /**
+          * Name of a wcs-mat-icon to display at the end (right) of the button.
+         */
+        "endIcon"?: string;
         /**
           * Set a URL to point to.<br/> If specified use a `a` tag instead of `btn`.
          */
@@ -5148,6 +5163,10 @@ declare namespace LocalJSX {
           * Specify the size of the button.
          */
         "size"?: WcsButtonSize;
+        /**
+          * Name of a wcs-mat-icon to display at the start (left) of the button.
+         */
+        "startIcon"?: string;
         /**
           * Specifies where to open the linked document when using href (see prop above)<br/> Default '_self' will open the linked document in the same frame as it was clicked
          */
@@ -7663,6 +7682,7 @@ declare module "@stencil/core" {
              * @cssprop --wcs-button-font-size-l - font-size for a size l button
              * @cssprop --wcs-button-padding-size-l - padding for a size l button
              * @cssprop --wcs-button-font-weight - font weight of a plain,stroked,clear button
+             * @cssprop --wcs-button-gap - gap between button content (text and icons)
              */
             "wcs-button": LocalJSX.WcsButton & JSXBase.HTMLAttributes<HTMLWcsButtonElement>;
             /**